kernel: Boost whenever a zygote-forked process becomes a top app
*Rework https://github.com/celtare21/kernel_essential_msm8998-1/commit/06e483835171a9a0e98accb2fc1ffcae482eaf13

Boost to the max for 500 ms whenever the top app changes, which
improves app launch speeds and addresses jitter when switching between
apps. A check to make sure that the top-app's parent is zygote ensures
that a user-facing app is indeed what's added to the top app task group,
since app processes are forked from zygote.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: celtare21 <celtare21@gmail.com>